Fetch and apply scores based on DOM modification event
Previously, fetching the data and applying changes required a long setTimeout to ensure the DOM was fully loaded.

Now, we are observing the DOM for changes, and debouncing the handler to ensure we only fetch and apply changes once. This also allows us to avoid the arbitrary 8000ms timeout, so the update is almost instantaneous. From manual testing, waiting 300ms after detecting the `match_` elements resulted in optimal performance (no missed race conditions).